requirements.txt
openai
langchain
langchain-community
chromadb
hydra-core
langchain-openai
docx2txt
networkx
pandas
numpy
unstructured
openpyxl